requirements.txt
# Librerías necesarias
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.2.1
sqlparse==0.5.3
tzdata==2025.2
qrcode[pil]==8.2

# Comando: pip install -r requirements.txt
# OJO: Ejecutar el comando cuando estés en la carpeta donde esté el archivo requirements.txt
